Parse the document's source string
Parse the source string in stead of the provided Source instances
because the graphql library doesn't work reliable when mixing different
versions.

This can happen when the user of FakeQL uses a different version of
graphql. It would provide us with a document Source instance created by
that version. We would then attempt to parse it with our version which
the graphql library does not support.

Specifically it breaks this check:
https://github.com/graphql/graphql-js/blob/157edcb4f7dfe9cb2acd3d82831430d983bbfbcf/src/language/parser.js#L172-L175

For a better explanation of what is happening here, take a look at
`isSchema`:
https://github.com/graphql/graphql-js/blob/157edcb4f7dfe9cb2acd3d82831430d983bbfbcf/src/type/schema.js#L50

It uses a custom `instanceOf` function which provides a better error:
https://github.com/graphql/graphql-js/blob/157edcb4f7dfe9cb2acd3d82831430d983bbfbcf/src/jsutils/instanceOf.js#L30-L41