Add support for customizing validation rules
It's possible that a document cannot be validated. This allows the user
to explicitly tell FakeQL what rules they expect to pass.

This can be useful, for example, with Apollo's @client directive. It's a
valid directive, but it won't exist in any schema. This allows users to
disable the validation for it.