Fix faking nested and sibling types
We appeared to have support for nested and sibling types but it was mere
luck. We didn't support queries like this for example:

query {
  friends {
    name
  }
  role
}

In the above query we would enter the friends list and never leave
again. Because role appears after friends, this meant we would assign
role to friends. The resulting mock was:

{
  friends: [
    {
      name: `mock-value-for-field-"name"`,
      role: `mock-value-for-field-"role"`
    },
  ],
}

Instead of:

{
  friends: [
    {
      name: `mock-value-for-field-"name"`,
    },
  ],
  role: `mock-value-for-field-"role"`,
}

Similarly fragments only appeared to work because they were never nested
and always the last to appear in the gql which meant the resulting
object just happened to be what we expect because the path to insert the
fragment at just happened to be correct. The tests were updated with
increasingly complex examples and they no longer pretend that we support
fragments.

The above is fixed by implementing a leave function which updates the
path when we leave nested or sibling fields.